+++ ECommerceApiTests/ProductTests.cs
@@ -66,7 +66,7 @@ namespace ECommerceApiTests
 
         [Theory]
         [InlineData("/product/create-product", "P1", 10, 100, HttpStatusCode.OK)]
-        public async Task Return_OK_When_Product_Created(string createProductUrl, string productCode, double price, double stock, HttpStatusCode expectedStatusCode)
+        public async Task Return_OK_When_Product_Created(string createProductUrl, string productCode, double price, int stock, HttpStatusCode expectedStatusCode)
         {
             var expectedResult = $"Product created; code {productCode}, price {price}, stock {stock}";
 
@@ -109,7 +109,7 @@ namespace ECommerceApiTests
 
         [Theory]
         [InlineData("/product/create-product", null, 10, 100, HttpStatusCode.BadRequest)]
-        public async Task Return_Bad_Request_Without_ProductCode(string createProductUrl, string productCode, double price, double stock, HttpStatusCode expectedStatusCode)
+        public async Task Return_Bad_Request_Without_ProductCode(string createProductUrl, string productCode, double price, int stock, HttpStatusCode expectedStatusCode)
         {
             // Arrange
             var request = new ProductInput
