+++ ECommerceCore/Managers/ProductManager.cs
@@ -7,7 +7,7 @@ namespace ECommerceCore.Managers
 {
     public class ProductManager
     {
-        public static string CreateProduct(string productCode, double price, double stock)
+        public static string CreateProduct(string productCode, double price, int stock)
         {
             IProduct product = new Product(productCode, stock, price);
             var existingProduct = Pool.Products.FirstOrDefault(p => p.Code.ToLower() == productCode.ToLower());
