+++ ECommerceCore/Objects/OrderItem.cs
@@ -5,7 +5,7 @@ namespace ECommerceCore.Objects
 {
     public class OrderItem
     {
-        public OrderItem(Guid productId, double productPrice, double quantity, List<Guid> appliedCampaigns)
+        public OrderItem(Guid productId, double productPrice, int quantity, List<Guid> appliedCampaigns)
         {
             Id = Guid.NewGuid();
             ProductId = productId;
@@ -16,7 +16,7 @@ namespace ECommerceCore.Objects
         public Guid Id { get; set; }
         public Guid ProductId { get; set; }
         public double ProductPrice { get; set; }
-        public double Quantity { get; set; }
+        public int Quantity { get; set; }
         public List<Guid> AppliedCampaignIdList { get; set; }
     }
 }
